+++ NEWS.md
@@ -78,3 +78,4 @@
 
 * Added negative binomial option to genCorGen, addCorGen, genCorFlex, and addCorFlex
 * Fixed bug in function genFactor
+* Added LAG() functionality to missing data generation - updated functions genMiss and added two new internal functions .checkLags and .addLags